rearranged load balancing classes
